Bump version to v1.2.5
Fixed an issue with Win32_DiskDrive where some of the drives would not be detected due to the names being their uniqueID. WMIC cannot recognize DeviceIDs as their UID, so this was temporarily used as a fallback option. However, the Index Property works just as fine and detects all the drives and can identify each of them accurately, even if there are duplicate names.

Win32_SoundDevice suffers from the same issue. However, there are no other unique identifiers besides DeviceID, which does not work.