Attempt 1/2
This attempt tries to rely on aliases provided for the actual Win32 class names. There might be a issue where the actual class names used in the main branch did not exist back when Windows 7 was released and hopes that the aliases point to the classes which existed at that time.

If this attempt is unsuccessful, I can try looking for the properties that exist in Windows 7 and capture only those values as my second attempt.

If all of my attempts fail, this branch will be DELETED and the repository will be ARCHIVED.

NOTE: There seems to be no alias for the class Win32_VideoController, which means no GPU information can be collected anymore if this attempt is met with success.